Update debugging.md (#355)
* Update debugging.md

1. Changing obsolete RenderObject method debugDescribeSettings to newer RenderParagraph method debugDescribeChildren.

This is a big of a guess, so please confirm.

2. Changing link to obsolete scheduler/Scheduler/addPersistentFrameCallback to scheduler/FrameCallback.

* Incorporating feedback

Addressing Hixie's feedback. For 1., the RenderObject debugFillDescription() is now in the material library, which is why it originally seemed suspect. LMK if the link should point to https://docs.flutter.io/flutter/rendering/Layer/debugFillDescription.html or https://docs.flutter.io/flutter/widgets/Widget/debugFillDescription.html instead.

* Changing debugFillDescription link to rendering/Layer version

Multiple classes in multiple libraries have a debugFillDescription method. The RenderObject class in the rendering library no longer has this method, but the Layer class does.  So, linking to this version at the end of the Rendering Layer section.