Adding new HTML/CSS analog page (#382)
* Adding new HTML/CSS analog page

Includes a new, wider layout (code_comp_page.html) with side-by-side code snippets. It inherits from default.html.

Also includes new two_column CSS file to format this type of content on a variety of devices. Content begins with <div class="begin-examples"> and ends with <div class="end-examples">. Enclose a new left-hand snippet in a <div> tag with CSS class "lefthighlight" and a new right-hand snippet in a <div> with class "righthighlight".

* Adding new CSS file for side-by-side code

* Adding new HTML/CSS analogs page

* Update home_sidebar.yml

* Incorporate @sfshaza comments

The CSS properties are not in code font because we want users to focus on the Flutter equivalents, and it's too font-noisy.
If this is breaking the style guide rules, LMK ;-)

* Replicate #387 MaterialList>ListView

Shameless avoidance of merging - replicating #387 in example, lines 637-661.
Code snippet failed TravisCI validation check because MaterialList is deprecated!

* Merging in #383 changes

* Merging in #383 changes

* Merging #384

* Removing MD2, ghost <title>

* Aligning @media w/bootstrap