Add pod setup step to iOS install instructions.
CocoaPods needs to download the master spec repo, which can take quite a
while. Adding a `pod setup` step to the installation instructions, so
users take the hit during install, instead of on the first build.

Moved CocoaPods installation steps to a separate section, since they are
needed for Simulator builds as well.

Fixes flutter/flutter#8983.