Move continuous integration to Cirrus from Travis (#1138)
This moves the CI setup to use Cirrus CI instead of Travis, to be in line with the other Flutter repos.

I've renamed some of the scripts so that they aren't Travis specific, and updated their bash to be more idiomatic and safe for filenames with spaces.

I've moved most of the install into a Docker image that can be updated with the tool/docker_linux/docker_build.sh and tool/docker_linux/docker_build.sh scripts. This includes the Ruby gems bundles, so the Gemfile and Gemfile.lock have moved to tool/docker_linux as well.

One unrelated change I had to make was to update two external URLs that were returning 404s in bootstrap-into-dart.md and widgets-intro.md. The build wouldn't pass without this change.

With the Docker image configuration, PR checks now take two and a half minutes to complete.