Rearrange the text on the Widget Inspector page (#1743)
I was very confused during my initial reading of the Widget
Inspector page; it starts off by saying "to start click ... on the
Flutter inspector toolbar..." without explaining where to find the
toolbar.  Meanwhile, the "Get Started with the inspector" section
is at the end of the page, and only there does it explain that the
inspector is available in the plugin for Android Studio or for
IntelliJ.

Rearrange the text to move most of it under the "Get Started"
section.  Also adjust the section name to have more typical tense
and capitalization.