Adding license file documentation
We current don't have a place for the documentation of the LICENSE file formatting rules. This PR contains an idea of how it can look like. The wording probably needs to be polished. 
We can also create a whole new page to have the documentation. I am not sure which way is better: put the documentation in place under the `Adding documentation` section or create a new page for it. 
cc @Hixie 

https://github.com/flutter/flutter/issues/31690