Minor correction to Flutter widget test cookbook
The headline for one of the steps in the widget test cookbook article is "Verify that the Widget uses a Matcher," which I don't believe is correct. What the code example shows is a Matcher being used to verify the behavior of a widget. I've tweaked the headline to reflect this.