Add instructions for updating the version number
A lot of Flutter users are publishing their apps to the Play store but finding a way to update the version number and name isn't straightforward. 

Hence, this answer is gaining more attention every week: https://stackoverflow.com/a/56970752/5066615

Even though the instructions are directly available in `pubspec.yaml`, most users (including me) cannot find it easily.  

Hence, this PR uses the same words from the `pubspec.yaml` on updating the version number of the app.