Suggestion for: Instructions on how to generate .pfx certificate for MSIX distribution.
MSIX lib currently lacking, is the way to build the .pfx certificate that is kind of equal to the .keystore file in Android apps for signing Release build.
Without a .pfx certificate, you can't install the windows app on a computer with MSIX Installer. And on Internet, it is all confusing and out-dated.

So, I have created a simplified step-by-step post to generate .pfx certificate.
https://www.reddit.com/r/FlutterDev/comments/lzo9cq/generate_a_pfx_certificate_for_flutter_windows/
Hope it can help flutter devs through the documentation.