Correct PATH modification instructions
Previously we asked users to modify their path as follows:

    export PATH="$PATH:[PATH_TO_FLUTTER_GIT_DIRECTORY]/flutter/bin"

The correct location is just `bin` relative to the main Flutter
directory, so the correct instructions should be:

    export PATH="$PATH:[PATH_TO_FLUTTER_GIT_DIRECTORY]/bin"

I've also made some minor wording changes to remove any ambiguity that
`PATH_TO_FLUTTER_GIT_DIRECTORY` is actually the path of the cloned
`flutter` repo and not the path into which the clone was landed.

This was a potential source of confusion/ambiguity since the initial
install instructions do suggest they run the following command to
temporarily add Flutter to their PATH after the initial install.

    export PATH="$PATH:`pwd`/flutter/bin"

Credit for spotting this error goes to christiannc on StackOverflow:
https://stackoverflow.com/questions/66713077/setting-up-flutter-in-terminal-yields-a-command-not-found-message/66716562#66716562