Make README.md symbolic link as per examples/
Two reasons for this PR -
1. To make the repo a bit lighter (9177 bytes) using symbolic link.
2. Running `tool/code-check.sh` and `tool/code-check.sh --null-safety`
hydrates the samples of `examples/` and `null_safety_examples/` if it
contains test folder.
`flutter create --no-overwrite $sample` creates the template README.md
file if not already present.
To avoid tracking of the newly created README.md, it is better to add
symbolic link to the `examples/` and `null_safety_examples/` folder's
README.md