Update note about web and plugins
The note at the top of the page about federated plugins is confusing, since federation and communication method are actually completely orthogonal, and this page isn't actually about plugins in the first place. This removes discussion of federation and plugins entirely, and instead talks about web implementation of platform-specific code directly.