Use package level import instead of direct path
The original typo fix continued to use a direct path ('../lib/main.dart') for the import. I didn't realize it at the time because the integrated test ran successfully, but this is discouraged by the compiler, and doing the same in my own project led to some confusing errors when running the tests. Changing the import to use a package level import seems to work in both cases, and is more appropriate.