Update codelab-web.md step 0 observations
This change adds the word "to" to where it seems to belong. Under the observations for Step 0, there's mention of how "the app's UI adheres Material Design" which doesn't sound quite right and would demonstrate an atypical use of the word 'adheres'. It seems more likely this is a typo.