Remove outdated video covering GetX (#8478)
I made this video almost 3 years ago about GetX package, and it has been
outdated for a while. I don't believe this is a good source to learn
GetX anymore. I haven't been keeping up with the development of the
project, and in general there hasn't been any updates to the package in
over 10 months.

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>