Update VS Code / DevTools section with latest info/screenshots (#8618)
This PR updates the text and screenshots for running DevTools in VS Code
to reflect the current stable behaviour. The existing text/screenshots
are from before DevTools was embedded, and before it was shipped inside
the SDK.

## Presubmit checklist
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.