[Proposal] Breakup development directory (#8624)
This pull request extracts all subcategories from `/development` to
top-level entries, to match similar entries like "Deployment" and
"Testing and debugging". The subcategories under Development are perhaps
the most important categories for learning Flutter, but they were hidden
under Development. This made them harder to navigate, with smaller text,
and with deeper links and breadcrumbs.

Work done:
- Pulled subdirectories out of `/development`
- Updated all old redirects and links to new destination
- Introduce new redirects so old links keep working
- Add some of the new top-level dividers to visually distinguish content
- Enable breadcrumbs in moved content
- Enable breadcrumbs within "Deployment"
- Moved "Add to app" below "Deployment"
- Add a short title for Add to app

This is part of incremental work, and will be followed up with breaking
up and reorganization "User interface", adjusting titles of content, and
adding some cookbooks to the sidenav.

Staged:
https://flutter-docs-prod--pr8624-feature-breakup-deve-00ees3e9.web.app/