Simplify 'Samples & tutorials' top-level category (#8807)
The Flutter Gallery shows a lot of the relevant code within the app and
has links to the source, so I think linking to the app should be
sufficient and reduce unnecessary choices.

As for the "Samples apps on GitHub", it leads to a standalone site, so
the "on GitHub" is a little confusing and a bit unnecessary. Especially
since not all of the demos are actually hosted on the flutter/samples
repo.