Correct asset variant documentation (#8710)
Fixes #8452, updating the [Adding assets and
images](https://docs.flutter.dev/ui/assets-and-images) page.

Because the idea of asset variants is limited to resolution-aware image
assets, this PR deletes the outdated [Asset variants
](https://docs.flutter.dev/ui/assets-and-images#asset-variants) section
and moves the relevant information to the documentation of
[resolution-aware image
assets](https://docs.flutter.dev/ui/assets-and-images#resolution-aware).

Other things this PR does:
* Moves information about the resolution-aware image asset feature to be
after the more fundamental information about how to load an image.
* Renames the "Declaring resolution-aware image assets" section to
"Resolution-aware image assets".
* Adds a subsection to this, called "Bundling of resolution-aware image
assets". This is to provide the information that the Asset variants
section provided. A subsection was chosen to match the existing scheme
of the [Asset images in package
dependencies](https://docs.flutter.dev/ui/assets-and-images#from-packages)
section.

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Anthony Sansone <atsansone@users.noreply.github.com>