Updated Happy Paths program pause notification (#8866)
Updated the notification that Happy Paths is paused. Removed the mention
that Flutter Favorites is paused as well since we are actively
restarting that program.

_Description of what this PR is changing or adding, and why:_ Updated
the notification that Happy Paths is paused. Removed the mention that
Flutter Favorites is paused as well since we are actively restarting
that program.

_Issues fixed by this PR (if any):_ N/A

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Anthony Sansone <atsansone@users.noreply.github.com>