Revert HTML changes from "Update template HTML with a11y fixes (#8819)" (#8946)
Reverts the HTML/template changes from #8819.

It introduced some issues, primarily not being able to close the side
menu on mobile/small screens.

There are a few minor formatting and semantic issues as well, such as
incorrect line breaking in the cookie notice, some less direct text in
the footer, multiple `<main>` elements on the 404 page, and not
following our style in a few cases.

To fix the issues quickly, this PR reverts the changes. Improvements
from the original PR can be revisited as follow-up, and should be done
isolated, rather than in one big PR.

This PR does not revert the JS changes from the PR as there are
conflicts due to other recent changes made and they seem to be working
fine.