Add categorical headers to Resources in sidenav (#8816)
The resources category was getting overwhelmingly large and hard to
visually parse, but dropdowns will hide a bunch of the content.

This aims to follow a similar design to devsite and others which have
non-dropdown headers. Open to feedback on the design, layout, etc. Happy
to drop this altogether as well if it seems like the wrong direction.