Incorporate networking cookbooks in sidenav with new tooling (#9061)
Introduces new tooling for sidenav active entry calculation (which entry
is open and should be highlighted). Previously it was one pass, so only
found the first entry which begins with the same path. So cookbook
entries found `/cookbook` under Samples and tutorials. The new
functionality generates a sort of tree structure which finds the closest
path but allows fallback to the next closest. So
`/cookbook/networking/fetch-data` will prefer that, then
`/cookbook/networking`, then `/cookbook`. This process is a little more
complicated, but determining activenav data for each page is much
faster, scaling with the path rather than how many sidenav entries there
are.

As a demo and first use of this functionality, add some of the
networking cookbooks under **Data & backend > Networking** in the
sidenav. It also adjusts their prev -> next order to better match a
potential learning workflow.

Also implements some validation of the `sidenav.yml` file and
regenerates all pages when the sidenav is updated.

---

Closes https://github.com/flutter/website/issues/9073, contributing to
https://github.com/flutter/website/issues/9013
Closes https://github.com/flutter/website/issues/1778 by adding
validation to permalink entries in the sidenav data