Fix and cleanup implicit animations codelab (#9201)
Essentially this PR is focused on improving the clarity of the steps
which are diff based, matching starting and ending runnable samples to
the steps, removing unnecessary code, and completing some cleanup of the
examples and the surrounding text.

**Detailed list of changes:**

- Sets up code excerpts for the starting and complete example code which
were mismatched with the actual steps/instructions. Then updates them to
match.
- Adds missing type specifiers and simplifies variable declarations.
- Adjusts trailing commas and formatting in opacity sample for easier to
read diffs.
- Removes extra folders/steps that are not currently used.
- Adds missing type specifiers and simplifies variable declarations.
- Moves away from inline ignores since the diff tooling doesn't remove
them, making the diffs more confusing and the line numbers inconsistent
between steps. Ignoring in the analysis options file allows for
consistent line numbers.
- Changes text of button from "change" to "Change" to match
instructions.
- Fixes a few instructions in the codelab's text.
- Removes the embedded survey.
- Adds a README to the example directory explaining its setup.

Closes https://github.com/flutter/website/issues/6929