[Windows] Document the show window migration (#9269)
Flutter 3.7 and 3.10 has a race condition that can result in the window
not being shown (see https://github.com/flutter/flutter/issues/119415).
Projects need to be migrated to fix this bug.

Most Windows projects will be migrated automatically by running `flutter
windows build` or `flutter run` unless they've done significant changes
to their `windows/runner/flutter_window.cpp` file.

This documents how to migrate a Windows project manually. Note that this
fix depends on an API that was added in Flutter 3.13. In other words,
projects that are still using Flutter 3.7 or 3.10 should not do this
migration.

_Issues fixed by this PR (if any):_
https://github.com/flutter/flutter/issues/127695


![image](https://github.com/flutter/website/assets/737941/a9ed95bf-91be-4a23-9857-fb888fee244a)