Improve Linux Install 1: Add links to Linux requirements in _linux-desktop-setup.md (#9081)
**_Description of what this PR is changing or adding, and why:**

This PR presents a comprehensive rewrite of the Flutter installation
instructions for Linux. The motivation behind this update is to make the
guide more accessible, clearer, and more reader-friendly for users new
to Flutter and Linux. This rewrite includes the addition of a glossary
for Linux terminal commands, step-by-step instructions on installing
Snap (if necessary), and restructured sections for improved flow and
readability.

**_Issues fixed by this PR (if any):**

No specific issues were addressed.

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Anthony Sansone <atsansone@users.noreply.github.com>