🔥 Remove `flutter` as a submodule (#9307)
_Description of what this PR is changing or adding, and why:_
The PR removes the `flutter` repo as a submodule to avoid conflicts when
building the website on different platforms. It was first introduced in
#1257 but with no explanation so we cannot track back what's the
purpose, but removing this should not cause difficulties in maintaining
the website.

### Additional info

- The build process is slower than the previous one because the checkout
step needs an extra 3~5 minutes, depending on the network condition. See
[now](https://github.com/flutter/website/actions/runs/5985413443/job/16237673762?pr=9307)
and
[previous](https://github.com/flutter/website/actions/runs/5983398408/job/16237673076?pr=9298)
checks.

_Issues fixed (partially) by this PR (if any):_
- https://github.com/flutter/website/issues/6201
- https://github.com/flutter/website/issues/6443
- https://github.com/flutter/website/issues/8272

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Anthony Sansone <atsansone@users.noreply.github.com>
Co-authored-by: Parker Lougheed <parlough@gmail.com>