Simplify test setup and fix inaccessible Dart/Flutter in container (#9416)
https://github.com/flutter/website/pull/9307 caused `dart` and `flutter`
to not be accessible when running inside the container. This PR attaches
the cloned Flutter repository as a volume to allow it to persist and be
accessible.

To be compatible with these changes, also modifies the `test.sh` script
to be simpler, not allowing switching branches separately from the
container. When it's ran from GitHub actions, a new container is created
anyway.