Rename Accessibility & localization to Accessibility & internationalization (#9507)
This PR updates the directory name from "Accessibility & Localization"
to "Accessibility & Internationalization" for better consistency with
the document titles within the directory.

Currently, the directory contains two documents: "Accessibility" and
"Internationalization". The term "Localization" in the directory name
does not accurately reflect the content of the "Internationalization"
document. By updating the directory name, we can ensure that both
document titles are represented consistently in the directory name.

With this change, developers browsing the documentation will have a
clearer understanding of the content covered in this directory.

---------

Co-authored-by: Brett Morgan <brettmorgan@google.com>
Co-authored-by: Parker Lougheed <parlough@gmail.com>