Correct Misleading Comment in the Code Snippet in background-parsing.md
In the final code snippet of section 3, the comment above the statement
`return parsePhotos(response.body);`  says,
`// Use the compute function to run parsePhotos in a separate isolate.`.
This comment is incorrect as we are not using the `compute` function here.
The `compute` function is introduced in section 4, where this exact
comment is repeated in a code snippet above the statement
`return compute(parsePhotos, response.body);`.

Hence, this comment should be reworded in the last code snippet of
section 3, where we have the statement `return parsePhotos(response.body);`.
I suggest it should be reworded as
`// Synchronously run parsePhotos in the main isolate.`
because this wording improves the clarity of where the statement below it,
executes.