Refactor the cloudbuild template (#9752)
NOTE: Given the current status of cloud build logs not publishing to
github reliably, this will not be enabled after this PR. The intention
of this PR is to clean up the current script to make it more
maintainable.

Refactors the current cloudbuild template for staging
* Moves the cloudbuild template to the `cloud_build` directory
* Adds a custom docker image that can be built and deployed to gcr.io
regularly, which includes `gh-cli` and `firebase-tools`. This abstracts
a lot of the "magic" of installing these tools into a docker image, and
also saves the time of having to reinstall them on the docker image
every single time this staging job runs.
* Refactors the current template by moving the logic to scripts, and
then call the scripts from the cloudbuild template. The scripts
themselves have also been organized for reability.