Updated commands for newer convention (#9909)
_Description:_

This PR updates the command for building Linux apps for distribution,
according to the new style of folder structure used by Flutter. The
previous command was outdated and did not work with the latest version
of Flutter.

_Issues fixed by this PR:_

Fixes #1234 

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.