Remove survey metadata redirect (#9954)
this redirect was accidentally committed.
related:
https://github.com/flutter/website/commit/919cb746459fb46a80ea69244557e0ffebf55388

The goal is to land https://github.com/flutter/website/pull/9942, but we
need to confirm the redirects are working even on master. I was noticing
some strange behavior there with trying to access the cloud storage urls
but I have not had time to dig into it.

This redirect is causing some failures on master so we need to land this
asap. Once this is green and approved, feel free to land as I'm going to
be away from my computer starting tomorrow.