Remove most Docker usage (#10216)
This PR results in most of the site infrastructure no longer being
functional when developing locally. This is a temporary step in the
transition away from Docker and Jekyll to make the final migration and
its change review easier. You'll need to temporarily rely on the testing
and staging on GitHub. Some testing can be ran locally with `dart run
flutter_site <command>` if you have the latest Flutter installation.

This PR also pulls out the code excerpt check, as it's independent of
the Dart channel and testing. This avoids unnecessary work and enables
quicker visibility into failure reasons on CI :)

Please let me know if you have any concerns with this landing or need
help with something after it lands. Thank you!

Contributes to https://github.com/flutter/website/issues/10203
Closes https://github.com/flutter/website/issues/9450