Add missing note about building Android plugin (#10362)
Every other platform's section had the note about needing to build at
least once before opening the native project, but it was missing for
Android; this adds it.

It also updates the iOS and macOS docs to add the `--config-only` flag,
which is an option that was added after these docs were originally
written, and makes the step much faster. (Windows and Linux are
unchanged since that option does not yet exist for those platforms).

_Description of what this PR is changing or adding, and why:_

_Issues fixed by this PR (if any):_

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.