Reorder duplicated Kotlin and Java snippets (#10375)
After discussing with others, it seems to make sense to place Kotlin
snippets before Java. It's been the Flutter tool's default for many
years and is now Android's default as well. The one exception is in the
add-to-app docs, where Java might be more common for old apps, but this
PR also puts Kotlin first there as well for consistency.

Closes https://github.com/flutter/website/issues/10364