Update index.md (#10382)
_Description of what this PR is changing or adding, and why:_
`Alignment.centerRight` throws off the alignment of the icon with the
background of the button on hover. When changing the alignment to
`Alignment.center`, the gap between the `IconButton` and
`_favoriteCount` is a bit bigger. Ultimately I think merging this change
is a matter of opinion and take no offense if we keep the existing
alignment.
NOTE: This was targeting Linux desktop, not mobile.

_Issues fixed by this PR (if any):_ #10381

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Brett Morgan <brettmorgan@google.com>