More fixes to adding a platform pages (#10402)
- "devtools" is a term that generally refers to the Dart/Flutter
DevTools app, we shouldn't conflate it with the various
apps/software/tools developers using when building for/developing
on/debugging against a platform. This PR continues and expand on changes
began in
https://github.com/flutter/website/commit/8c0003e94988362179fd8532e201439425ed5ecd
to move away from these usages of "devtools".
- The
[web](https://docs.flutter.dev/platform-integration/web/install-web) and
[android](https://docs.flutter.dev/platform-integration/android/install-android)
pages had some rendering issues due to whitespace causing rendering
differences. This PR addresses those.
- This PR also standardize styles and Liquid usage between the
equivalent pages for different platforms.

Feel free to push any changes you'd like to this branch. I haven't had a
chance to understand what content is covered or to put too much thought
into the names. I put the page names as "Add X as a target platform ..."
for now to be consistent with
https://github.com/flutter/website/commit/8c0003e94988362179fd8532e201439425ed5ecd,
but it seems the content might actually be "Set up your development
environment for targeting X platform with Flutter"?

Closes https://github.com/flutter/website/issues/10389

@atsansone @sfshaza2 I've having trouble wrapping my head around these
pages, or perhaps just their structure. It's also not clear to me what
developers who didn't or don't follow the current version of the getting
started instructions should do. I think it might be beneficial if us
three could chat through this whole flow together. For now though, I
kept this individual PR focused on fixing the rendering issues and
replacing the "devtools" mentions.