Update pub add.md
Updated the instructions on how to oub add dev dependencies for integration tests. 

Made more brief. 
I separated the command from the console output. In order to make easy to cut and paste command. And to condense output.

I included what the updated lines of the pubspec should  look like