Revert "Add a link to the Flutter for web developers guide" (#10428)
Reverts flutter/website#10422

The content at `/get-started/install` is not really an introduction to
Flutter for web developers. It's how to get started using Flutter for
the web.

I'd argue we probably need to think about what content we would give a
developer coming from the web to Flutter, but "install these tools" is
probably not it.