more on Flutter web indexability (#10550)
Expand on what we're doing about indexability.

One unusual aspect of this change is linking to the test site and
spelling out "Hawaii-themed space stories". This is intentional, and
hopefully, not too disruptive to the message. In order to prove that SEO
works we need an already indexed website to link to the test site so
search engines find it and attempt to index it. I thought sneaking one
such link into a section of our web FAQ that already talks about SEO
would be a safe place for it.

One place I violate the style guide is I use "we" in the text. However,
it seems to be consistent with how the word is used elsewhere on the
site. I'm open to suggestions for how to remove it, if it's important.

## Presubmit checklist

- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.