minor: change transformers args to list (#10622)
The `args` in the `asset-transformation` section should be of type list
of strings, not a single string as the documentation shows. Running the
code as in the docs will result in the following issue:

```
Error detected in pubspec.yaml:
Unable to parse assets section.
In transformers section of asset "assets/logo.svg": In args section of transformer using package "vector_graphics_compiler": Expected args to be a list of String, but got --tessellate --font-size=14 (String).
Please correct the pubspec.yaml file
```

This PR convert the args section in the docs to list.